move 'Button' component to 'src/components' folder
