+++ src/modules/pages/ApiListPage/ApiListPage.js
@@ -8,7 +8,7 @@ import ApiListContainer from './apiListContainer';
 import Section from '../../Layout/Section/Section';
 import Row from '../../Layout/Row/Row';
 import Title from '../../Layout/Title/Title';
-import Button from '../../buttons/Button';
+import Button from '../../../components/Button/Button';
 import UnhealthyReport from '../HealthCheckPage/UnhealthyReport';
 
 const ApiListPage = ({ healthcheckStatus }) => {
