+++ src/modules/pages/OAuthServersPage/OAuthServersPage.js
@@ -7,7 +7,7 @@ import OAuthServersContainer from './oAuthServersContainer';
 import Section from '../../Layout/Section/Section';
 import Row from '../../Layout/Row/Row';
 import Title from '../../Layout/Title/Title';
-import Button from '../../buttons/Button';
+import Button from '../../../components/Button/Button';
 
 const OAuthServersPage = () => {
     return (
